moved LottoCoupon into lotto package
